[ABI] Added Fill/Stroke API with Style, improved rendering context
Core changes:

  * Changed the layout of BLObject metadata (removing 'I' flag, changing how
    object type is stored, and how object reference counting is checked)
  * Made it possible to free a typed BLObjectImpl without BLObjectInfo metadata
  * Moved some Impl properties stored in BLBobjectDetail to BLObjectImplHeader
  * Moved tests to separate files having _test.cpp suffix
  * Added Rgba32 and Rgba64 types to Blend2D object type system
  * BLBitSet now uses at most 2 BitWords in SSO Dense representation so it
    doesn't need special treatment of its ObjectType anymore
  * Refactored Blend2D SIMD abstraction layer, which now provides X86 SIMD up
    to AVX-512 and ARM SIMD targeting both 32-bit and 64-bit platforms
  * Added initial SIMD tests to make it possible to verify the implementation

Rendering context changes:

  * Refactored rendering context frontend and dispatch mechanism
  * Reworked aligned and fractional pattern vertical extends (pipelines)
  * Removed BoxU filler (masks are always used to render BoxU commands)
  * Added Fill/Stroke functions that accept an explicit color or Style
  * Added dithering quality to gradients and implemented it in all fetchers
  * Added mask filler to portable pipeline to make it complete
  * Added 256/512-bit SIMD support to Conic gradient fetchers
  * Improved the performance of rendering conic gradients by tweaking how
    it transforms the coordinates
  * Improved support of partial fetches

+++ src/blend2d/support/lookuptable_p.h
@@ -76,21 +76,21 @@ struct BitTableGeneratorAdapter {
   }
 
   static constexpr BLBitWord gen32(size_t index) noexcept {
-    return (gen8(index +  0u) <<  0u) |
-           (gen8(index +  8u) <<  8u) |
-           (gen8(index + 16u) << 16u) |
-           (gen8(index + 24u) << 24u) ;
+    return (uint32_t(gen8(index +  0u)) <<  0u) |
+           (uint32_t(gen8(index +  8u)) <<  8u) |
+           (uint32_t(gen8(index + 16u)) << 16u) |
+           (uint32_t(gen8(index + 24u)) << 24u) ;
   }
 
   static constexpr BLBitWord gen64(size_t index) noexcept {
-    return (gen8(index +  0u) <<  0u) |
-           (gen8(index +  8u) <<  8u) |
-           (gen8(index + 16u) << 16u) |
-           (gen8(index + 24u) << 24u) |
-           (gen8(index + 32u) << 32u) |
-           (gen8(index + 40u) << 40u) |
-           (gen8(index + 48u) << 48u) |
-           (gen8(index + 56u) << 56u) ;
+    return (uint64_t(gen8(index +  0u)) <<  0u) |
+           (uint64_t(gen8(index +  8u)) <<  8u) |
+           (uint64_t(gen8(index + 16u)) << 16u) |
+           (uint64_t(gen8(index + 24u)) << 24u) |
+           (uint64_t(gen8(index + 32u)) << 32u) |
+           (uint64_t(gen8(index + 40u)) << 40u) |
+           (uint64_t(gen8(index + 48u)) << 48u) |
+           (uint64_t(gen8(index + 56u)) << 56u) ;
   }
 
   static constexpr BLBitWord value(size_t index) noexcept {
